chore: make CI gates runnable locally and check the whole repo for link rot
Ports the developer-experience hardening from ollygarden/humus (PRs #98-#107)
so a local run and a green build cannot disagree.

Every gate CI enforces is now a script in bin/, runnable from any directory:

- bin/validate-skill.sh delegates spec conformance to skills-ref and adds the
  house rule that a SKILL.md stays under 500 lines.
- bin/check-skill-inventory.py (moved from .github/scripts/) keeps skills/,
  marketplace.json, and the README table and tree in sync.
- bin/skills-ref.requirement holds the pinned validator revision. It was
  previously duplicated in ci.yml and CONTRIBUTING.md with nothing keeping the
  two equal; renovate.json now watches the single copy.

Two defects in the inventory checker, found in cross-repo review and fixed here
with reproduced failure cases:

- The README checks ran in one direction only, so a table row or tree entry
  naming a deleted skill was never reported and the README could advertise a
  skill that is not shipped. Both indexes are now set-diffed both ways, as the
  marketplace already was.
- The Repository Structure check was a bare substring match, satisfied by any
  indented listing anywhere in the file. It is now anchored and scoped to the
  skills/ block.

Link checking moves to its own workflow and covers the whole repository on
every PR plus weekly, rather than checking a single generated file. Link rot is
an external-world failure, so it does not belong in the required content check.
Every exclusion in lychee.toml is categorized with the reason it is not a real
link. Two genuine 404s surfaced and are fixed rather than excluded: the
anthropics/skills skill-creator path moved, and the Drain paper changed hosts.

Also adds a .gitignore, which the repository did not have. local/ was ignored
only via a maintainer's global gitignore, so scratch and research notes were
one `git add -A` away from a public repo while AGENTS.md asserted they were
safe. docs/preferred-workflow.md records how a change moves through the repo,
and AGENTS.md now marks each constraint as automated or review-enforced.
