+++ AGENTS.md
@@ -8,6 +8,30 @@ A catalog of **Agent Skills** (per the [agentskills.io](https://agentskills.io/s
 
 Most changes are to Markdown and YAML files that AI agents consume. The exception is `tools/otel-agent-tools/`, a Go module that generates and validates some of the bundled reference data (see below). "Correctness" means a skill is well-scoped, accurate, points at the maintained source of truth, and is registered in all the right places.
 
+## Preferred workflow
+
+For any change that adds, renames, moves, or removes a skill, or that alters what a skill triggers on or recommends, follow [`docs/preferred-workflow.md`](docs/preferred-workflow.md). Typo and prose fixes go straight to a PR — unless they touch `SKILL.md` frontmatter, since a `description:` edit changes when the skill activates.
+
+## Running the gates locally
+
+Everything CI enforces is a script in `bin/`, runnable from any directory:
+
+```bash
+./bin/validate-skill.sh              # spec conformance + house rules; a path checks one skill
+./bin/check-skill-inventory.py     # skills/, marketplace.json, and README in sync
+```
+
+`validate-skill.sh` needs `skills-ref`, pinned in `bin/skills-ref.requirement` — that file is the single source CI, Renovate, and `CONTRIBUTING.md` all read, so never paste a revision anywhere else. Install with `uv tool install "$(cat bin/skills-ref.requirement)"`.
+
+## Skill constraints
+
+- The directory name equals the `name:` field — **automated** by `bin/validate-skill.sh` (via `skills-ref`).
+- Frontmatter parses as YAML, `description` fits in 1024 characters, no unknown keys — **automated**, same script.
+- `SKILL.md` stays under 500 lines — **automated**. Past that, move detail into `references/`.
+- Registration across `skills/`, `marketplace.json`, and `README.md` — **automated** by `bin/check-skill-inventory.py`.
+- Links resolve — **automated** by the `Link Check` workflow, which scans the whole repository weekly and on every PR.
+- Keeping content vendor-neutral, DRY, and token-efficient — **review-enforced**, not automated. A green build says nothing about it.
+
 ## Architecture
 
 Each skill is a self-contained directory under `skills/<skill-name>/`:
@@ -43,7 +67,7 @@ A new skill is only "registered" when it appears in **all** of these. Missing an
 
 - **Agent-authored PRs are accepted** and expected — but a human must own the PR, and agent involvement should be disclosed in the description.
 - **Harness evidence is required** for any PR that adds or substantively changes a skill: run the same representative prompt(s) on a frontier model without and with the skill (fresh sessions, same model and harness), and include the comparison plus transcript links in the PR description. The `.github/PULL_REQUEST_TEMPLATE.md` has a section for this.
-- **Spec conformance**: validate with `skills-ref validate skills/<skill-name>` ([agentskills.io spec](https://agentskills.io/specification)).
+- **Spec conformance**: validate with `./bin/validate-skill.sh` ([agentskills.io spec](https://agentskills.io/specification)).
 - **CLA**: first-time contributors sign the organization-wide
   [OllyGarden CLA](https://github.com/ollygarden/.github/blob/main/CLA.md) via the CLA bot on the PR
   (`.github/workflows/cla.yml`).
@@ -55,4 +79,4 @@ A new skill is only "registered" when it appears in **all** of these. Missing an
 - **Design for token efficiency.** Avoid dumping large files or broad context when a targeted lookup, focused reference, or small generated artifact will do.
 - **Stay vendor neutral and non-opinionated.** Opinions belong in the companion `skills` repo.
 - A skill `description` is the trigger surface: it should enumerate concrete user phrasings so agents activate it reliably. Mirror the existing skills' description style.
-- `local/` is gitignored — used for scratch/research notes, never published.
+- `local/` is gitignored (see `.gitignore`) — used for scratch/research notes, never published. The link checker honours `.gitignore`, so scratch files cannot redden CI either.
