+++ bin/skills-ref.requirement
@@ -1,0 +1,1 @@
+skills-ref @ git+https://github.com/agentskills/agentskills.git@38a2ff82958afee88dadf4831509e6f7e9d8ef4e#subdirectory=skills-ref
