+++ renovate.json
@@ -1,0 +1,21 @@
+{
+  "$schema": "https://docs.renovatebot.com/renovate-schema.json",
+  "extends": [
+    "config:recommended"
+  ],
+  "customManagers": [
+    {
+      "customType": "regex",
+      "description": "Keep the pinned skills-ref revision fresh. The pin is a PEP 508 direct reference, which the pip_requirements manager skips, and the filename does not match its default pattern either — without this, the validator would silently drift behind the spec it enforces.",
+      "managerFilePatterns": [
+        "/^bin/skills-ref\\.requirement$/"
+      ],
+      "matchStrings": [
+        "git\\+https://github\\.com/(?<depName>[^/]+/[^/@]+?)\\.git@(?<currentDigest>[0-9a-f]{40})"
+      ],
+      "datasourceTemplate": "git-refs",
+      "packageNameTemplate": "https://github.com/{{{depName}}}",
+      "currentValueTemplate": "main"
+    }
+  ]
+}
